Conflicts:
	camp/camp-brooklyn/src/test/java/org/apache/brooklyn/camp/brooklyn/ConfigParametersYamlTest.java
	core/src/main/java/org/apache/brooklyn/core/config/internal/AbstractConfigMapImpl.java
	core/src/main/java/org/apache/brooklyn/core/internal/BrooklynPropertiesImpl.java
	core/src/main/java/org/apache/brooklyn/core/mgmt/internal/DeferredBrooklynProperties.java

Mostly `@Overrides` annotations added upstream, some tidying esp in BrooklynProperties for synchronization and removing unneeded methods

+++ core/src/main/java/org/apache/brooklyn/location/byon/ByonLocationResolver.java
@@ -102,7 +102,7 @@ public class ByonLocationResolver extends AbstractLocationResolver {
         Object hosts = config.getStringKey("hosts");
         config.remove("hosts");
         String user = (String) config.getStringKey("user");
-        Integer port = (Integer) TypeCoercions.coerce(config.getStringKey("port"), Integer.class);
+        Integer port = TypeCoercions.coerce(config.getStringKey("port"), Integer.class);
         Class<? extends MachineLocation> locationClass = getLocationClass(config.get(OS_FAMILY));
 
         MutableMap<String, Object> defaultProps = MutableMap.of();
